+++ game/src/api/Achievement.hpp
@@ -10,8 +10,8 @@ std::string toString() const;
 void fromString(const std::string &str);
         std::string id;
         int64_t progress;
-        std::optional<double> rd;
-        std::optional<double> rn;
+        std::optional<int64_t> rd;
+        std::optional<int64_t> rn;
         std::string unlocked;
     };
 }
